second pass

+++ libopus/silk_init_decoder.go
@@ -1,6 +1,6 @@
 package libopus
 
-func silk_init_decoder(psDec *silk_decoder_state) int64 {
+func silk_init_decoder(psDec *silk_decoder_state) int {
 	*psDec = silk_decoder_state{}
 	psDec.First_frame_after_reset = 1
 	psDec.Prev_gain_Q16 = 65536
